Reorganize serialization methods
