+++ lemon/__init__.py
@@ -1,6 +1,6 @@
 from .header import Header
-from .markdown import Markdown, dumps
-from .parse import loads
+from .markdown import Markdown
+from .serialize import dumps, loads
 from .style import Bold, Italics, Strikethrough
 from .tables import Table
